policy-schedule-reader/requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
argcomplete==3.6.0
asttokens==3.0.0
boto3==1.37.17
botocore==1.37.17
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cohere==5.14.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
eval_type_backport==0.2.2
executing==2.2.0
fastavro==1.10.0
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
google-auth==2.38.0
google-genai==1.5.0
griffe==1.6.2
groq==0.20.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.9.0
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
logfire-api==3.9.0
markdown-it-py==3.0.0
matplotlib==3.10.1
matplotlib-inline==0.1.7
mcp==1.4.1
mdurl==0.1.2
mistralai==1.6.0
nest-asyncio==1.6.0
numpy==2.2.4
openai==1.68.0
opentelemetry-api==1.31.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-ai==0.0.43
pydantic-ai-slim==0.0.43
pydantic-graph==0.0.43
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
requests==2.32.3
rich==13.9.4
rsa==4.9
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.46.1
tokenizers==0.21.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-requests==2.32.0.20250306
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websockets==14.2
wrapt==1.17.2
zipp==3.21.0
